requirements.txt
rdflib==7.0.0
requests
beautifulsoup4
openai
SPARQLWrapper
python-dotenv